requirements.txt
accelerate>=0.16.0
torchvision
transformers == 4.31.0
datasets
ftfy
tensorboard
Jinja2

